use UniGLTF.IBytesBuffer
