+++ Assets/UniGLTF/Runtime/UniGLTF/IO/ArrayByteBuffer.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 1e9c9201942704b4f9dd050115073032
+guid: d095d4c4eb5e68f4bb10c21bb3604cdb
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
