+++ Assets/UniGLTF/Runtime/UniGLTF/IO/UriByteBuffer.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: c700d6d6fbb38fa4d9662213c58d5725
+guid: 8b75aca0e4a7415418cd1c3d018483b2
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
